pulled functions out of bot and created separate libraries
ideally spotify.js will handle communications with the spotify api, commands.js and voting.js will handle discord messages, and data.js will handle reading and writing to files, and databases (once implemented).

A lot still needs to be unwound to reach this goal but this is a so far successful start with the app still functioning with the new changes.